+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added dependency to `keyple-plugin-storagecard-java-api:1.0.0` for support of storage cards
+### Build
+- Removed dependency to `keyple-gradle` plugin
+### Upgraded
+- Dokka to `2.0.0`
 
 ## [3.0.0] - 2025-02-20
 :warning: Major version!